example_d/requirements.txt
functions-framework==3.*
google-cloud-bigquery
earthengine-api>=0.1.358
